+++ src/cli/node/upgrade/upgrade.js
@@ -21,11 +21,12 @@ const {
   updateChainConfig,
   updateReleaseYaml,
 } = require('core/forge-fs');
+const { getChainVersion } = require('core/libs/common');
 const { isForgeStartedByStarter } = require('core/forge-process');
 const debug = require('core/debug')('upgrade');
+const { ensureModerator } = require('core/moderator');
 
 const { stop, waitUntilStopped } = require('../stop/stop');
-const { ensureModerator } = require('../../protocol/deploy/deploy');
 
 function isStoppedToUpgrade(chainName) {
   return new Promise(resolve => {
@@ -138,7 +139,10 @@ async function main({ args: [chainName = process.env.FORGE_CURRENT_CHAIN] }) {
       process.exit(0);
     }
 
-    const moderator = await ensureModerator(client);
+    const currentVersion = getChainVersion(chainName);
+    const moderator = await ensureModerator(client, {
+      currentVersion,
+    });
     if (!moderator) {
       return;
     }