+++ src/cli/protocol/deploy/deploy.js
@@ -1,115 +1,36 @@
 const path = require('path');
-const shell = require('shelljs');
 const chalk = require('chalk');
 const base64 = require('base64-url');
-const semver = require('semver');
 const { toItxAddress } = require('@arcblock/did-util');
-const { fromSecretKey } = require('@arcblock/forge-wallet');
-const { bytesToHex, hexToBytes, isHexStrict } = require('@arcblock/forge-util');
-const { symbols } = require('core/ui');
-const { printError, printInfo, sleep } = require('core/util');
+const { bytesToHex } = require('@arcblock/forge-util');
+const { messages } = require('@arcblock/forge-proto');
+const { print, printError, printInfo, printSuccess, sleep } = require('core/util');
 const { isFile } = require('core/forge-fs');
-const { config, createRpcClient } = require('core/env');
+const { createRpcClient } = require('core/env');
+const { getChainVersion } = require('core/libs/common');
+const { ensureModerator } = require('core/moderator');
+const { pretty } = require('core/ui');
 const debug = require('core/debug')('deploy');
 
-function ensureModeratorSecretKey() {
-  const sk = process.env.FORGE_MODERATOR_SK;
-  if (!sk) {
-    shell.echo(`${symbols.error} please set FORGE_MODERATOR_SK to continue`);
-    process.exit(1);
-  }
+const { fetchProtocols } = require('../list/list');
 
-  if (isHexStrict(sk)) {
-    return sk;
-  }
-
-  // debug('detected base64 moderator sk', base64.unescape(sk));
-  return bytesToHex(Buffer.from(base64.unescape(sk), 'base64'));
-}
-
-function ensureModeratorDeclared(client, wallet) {
-  return new Promise((resolve, reject) => {
-    const stream = client.getAccountState({ address: wallet.toAddress() });
-    let account = null;
-    stream.on('data', ({ code, state }) => {
-      if (code === 0 && state) {
-        account = state;
-      }
-    });
-    stream.on('end', async () => {
-      if (account) {
-        resolve(account);
-      } else {
-        const hash = await client.sendDeclareTx({
-          tx: {
-            itx: { moniker: 'moderator' },
-          },
-          wallet,
-        });
-        shell.echo(`${symbols.info} moderator declared ${hash}`);
-        resolve(hash);
-      }
-    });
-
-    stream.on('error', reject);
-  });
-}
-
-// eslint-disable-next-line consistent-return
-const ensureModerator = async client => {
-  const sk = ensureModeratorSecretKey();
-  const moderator = fromSecretKey(sk);
-  printInfo(`moderator address ${moderator.toAddress()}`);
-
-  let moderatorKey = 'forge.prime.moderator';
-  const currentVersion = config.get('cli.currentVersion');
-
-  if (semver.lt(currentVersion, '0.38.0')) {
-    moderatorKey = 'forge.moderator';
-  }
-
-  if (!config.get(`${moderatorKey}.address`)) {
-    printError('Abort because forge.moderator is not set in config file');
-    printInfo(
-      `Please add following content in config file ${chalk.cyan(
-        config.get('cli.forgeConfigPath')
-      )},`
-    );
-    printInfo(`${chalk.red('then restart current forge:')}`);
-    shell.echo(`
-[${moderatorKey}]
-address = "${moderator.toAddress()}"
-publicKey = "${base64.escape(base64.encode(hexToBytes(moderator.publicKey)))}"
-`);
-
-    process.exit(1);
-  }
-
-  try {
-    await ensureModeratorDeclared(client, moderator);
-    shell.echo(`${symbols.success} moderator declared on chain`);
-    return moderator;
-  } catch (err) {
-    shell.echo(`${symbols.error} ${err.message}`);
-  }
-};
-
-async function main({ args: [itxPath] }) {
+async function main({ args: [itxPath], opts: { chainName } }) {
   try {
     const itxFile = path.resolve(itxPath);
     if (!isFile(itxFile)) {
-      shell.echo(`${symbols.error} itx.json file ${itxFile} not exists`);
+      printError(`itx.json file ${itxFile} not exists`);
       process.exit(1);
     }
 
     const client = createRpcClient();
-    const moderator = await ensureModerator(client);
+    const currentVersion = getChainVersion(chainName);
+    const moderator = await ensureModerator(client, { currentVersion });
     if (!moderator) {
       return;
     }
 
     // eslint-disable-next-line no-underscore-dangle
-    shell.echo(`${symbols.info} deploy protocol to ${client._endpoint}`);
+    printInfo(`deploy protocol to ${client._endpoint}`);
 
     // eslint-disable-next-line
     const json = require(itxFile);
@@ -126,7 +47,21 @@ async function main({ args: [itxPath] }) {
     const DeployProtocolTx = client.getType('DeployProtocolTx');
     const itxObj = DeployProtocolTx.deserializeBinary(itxBuffer).toObject();
     itxObj.address = toItxAddress(itxObj, 'DeployProtocolTx');
-    shell.echo('transaction protocol detail', itxObj);
+
+    const protocols = await fetchProtocols(client);
+    const protocolOnChain = protocols.find(
+      ({ name, version }) => itxObj.name === name && itxObj.version === version
+    );
+
+    if (protocolOnChain) {
+      protocolOnChain.status = messages.ProtocolStatus[protocolOnChain.status].toLowerCase();
+      printError('Deploy failed: the protocol has been deployed.');
+      printInfo('Protocol info:');
+      print(pretty(protocolOnChain));
+      process.exit(1);
+    }
+
+    print('transaction protocol detail', itxObj);
 
     const hash = await client.sendDeployProtocolTx({
       tx: {
@@ -135,15 +70,14 @@ async function main({ args: [itxPath] }) {
       },
       wallet: moderator,
     });
-    shell.echo(`${symbols.success} transaction protocol deploy success`);
+    printSuccess('transaction protocol deploy success');
     await sleep(5000);
-    shell.echo(`${symbols.info} inspect tx with ${chalk.cyan(`forge tx ${hash}`)}`);
+    printInfo(`inspect tx with ${chalk.cyan(`forge tx ${hash}`)}`);
   } catch (err) {
     debug.error(err);
-    shell.echo(`${symbols.error} transaction protocol deploy failed: ${err.message}`);
+    printError(`transaction protocol deploy failed: ${err.message}`);
   }
 }
 
 exports.run = main;
 exports.execute = main;
-exports.ensureModerator = ensureModerator;