+++ src/core/moderator.js
@@ -1,12 +1,19 @@
 const base64 = require('base64-url');
+const get = require('lodash/get');
+const semver = require('semver');
+const chalk = require('chalk');
+const TOML = require('@iarna/toml');
 const { fromSecretKey } = require('@arcblock/forge-wallet');
 const { bytesToHex, hexToBytes, isHexStrict } = require('@arcblock/forge-util');
+
+const { print, printError, printInfo, printSuccess } = require('core/util');
+const debug = require('core/debug')('moderator');
 
 const { getGlobalConfig } = require('./libs/global-config');
 
 function getModeratorSecretKey() {
   const { moderatorSecretKey } = getGlobalConfig();
-  const sk = moderatorSecretKey || process.env.FORGE_MODERATOR_SK;
+  const sk = process.env.FORGE_MODERATOR_SK || moderatorSecretKey;
 
   if (!sk) {
     return undefined;
@@ -21,19 +28,76 @@ function getModeratorSecretKey() {
 
 function getModerator() {
   const sk = getModeratorSecretKey();
-  if (sk) {
-    const wallet = fromSecretKey(sk);
-    const pk = base64.escape(base64.encode(hexToBytes(wallet.publicKey)));
-    return {
-      address: wallet.toAddress(),
-      pk,
-      publicKey: pk, // TODO: removed?
-    };
+  if (!sk) {
+    printInfo('moderator sk was not set');
+    return undefined;
   }
 
-  return undefined;
+  const wallet = fromSecretKey(sk);
+  const pk = base64.escape(base64.encode(hexToBytes(wallet.publicKey)));
+  return {
+    address: wallet.toAddress(),
+    pk,
+    publicKey: pk, // TODO: removed?
+  };
 }
 
+const getModeratorFromChain = async (client, currentVersion) => {
+  if (semver.lt(currentVersion, '0.38.0')) {
+    const res = await client.getConfig();
+    const configString = get(res.$format(), 'config');
+    const config = TOML.parse(configString);
+
+    return get(config, 'forge.moderator', null);
+  }
+
+  const res = await client.getForgeState();
+  return get(res.$format(), 'state.accountConfig.moderator', null);
+};
+
+// eslint-disable-next-line consistent-return
+const ensureModerator = async (client, { currentVersion }) => {
+  debug('ensureModerator:currentVersion', currentVersion);
+
+  const sk = getModeratorSecretKey();
+  if (!sk) {
+    printError('The moderator secret key not found in your environment.');
+    print();
+    printInfo('The moderator secret key can be set in either of the following ways:');
+    print(`  1. set ${chalk.cyan('moderatorSecretKey')} in ${chalk.cyan('~/.forgerc.yml')}:`);
+    print('    moderatorSecretKey: xxx');
+    print(
+      `  2. set ${chalk.cyan('FORGE_MODERATOR_SK')} in ${chalk.cyan('~/.bashrc')} or ${chalk.cyan(
+        '~/.zshrc'
+      )}:`
+    );
+    print('    export FORGE_MODERATOR_SK=xxx');
+    print();
+    process.exit(1);
+  }
+
+  const chainModerator = await getModeratorFromChain(client, currentVersion);
+  debug('moderator on chain', chainModerator);
+
+  if (!chainModerator) {
+    printError('There is no moderator on the chain, so the operation can\'t be done.'); // prettier-ignore
+    process.exit(1);
+  }
+
+  const localModerator = fromSecretKey(sk);
+  if (localModerator.toAddress() !== chainModerator.address) {
+    printError('Abort because moderator address from your env does not match the moderator in chain state'); // prettier-ignore
+    print(`  - Local moderator: ${chalk.cyan(localModerator.toAddress())}`);
+    print(`  - Remote moderator: ${chalk.cyan(chainModerator.address)}`);
+    printInfo('To continue upgrade, please set the moderator secret key that have the same address in chain state in your env'); // prettier-ignore
+    process.exit(1);
+  }
+
+  printSuccess('moderator checked success');
+  return localModerator;
+};
+
 module.exports = {
+  ensureModerator,
   getModerator,
 };